+++ src/idc/dataformatter/__init__.py
@@ -1,0 +1,1 @@
+from ._image_data_formatter import ImageDataFormatter