+++ CHANGES.rst
@@ -35,6 +35,7 @@ Changelog
 - `idc-convert` and `idc-exec` now support the default placeholders HOME/CWD/TMP when loading custom placeholders,
   now load the placeholders correctly (fixed incorrect late lambda binding)
 - added `exif-autorotate` filter to automatically apply EXIF rotation to images (if applicable)
+- added support for data-formatters with `df-image-data` specifically for images
 
 
 0.1.0 (2025-10-31)